+++ src/main/java/lotto/exception/ErrorType.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.exception;
 
 public enum ErrorType {
